pass build. comment out context relative headers including, add missing directory function in pal-win32, etc.

+++ pal/win32/src/gui.cpp
@@ -256,8 +256,8 @@ xsBool xsGetClientRect(xsRect *rect)
 
 	rect->left = 0;
 	rect->top = 0;
-	rect->bottom = g_nScreenHeight - 1;
-	rect->right = g_nScreenWidth - 1;
+	rect->bottom = (float)g_nScreenHeight - 1;
+	rect->right = (float)g_nScreenWidth - 1;
 
 	return XS_TRUE;
 }
@@ -265,10 +265,10 @@ xsBool xsGetClientRect(xsRect *rect)
 void xsSetClipRect(xsGraphics *gc, xsRect *rect)
 {
 	Rect rc;
-	rc.X = rect->left + gc->xoffset;
-	rc.Y = rect->top + gc->yoffset;
-	rc.Width = rect->right - rect->left + 1;
-	rc.Height = rect->bottom - rect->top + 1;
+	rc.X = (INT)(rect->left + gc->xoffset);
+	rc.Y = (INT)(rect->top + gc->yoffset);
+	rc.Width = (INT)(rect->right - rect->left + 1);
+	rc.Height = (INT)(rect->bottom - rect->top + 1);
 	g_gcBase->SetClip(rc);
 }
 
@@ -313,12 +313,12 @@ void xsFillTriangle(xsGraphics *gc, float x1, float y1, float x2, float y2, floa
 {
 	Point points[3];
 
-	points[0].X = x1 + gc->xoffset - 1; // buggy GDI+?
-	points[0].Y = y1 + gc->yoffset - 1; // buggy GDI+?
-	points[1].X = x2 + gc->xoffset;
-	points[1].Y = y2 + gc->yoffset;
-	points[2].X = x3 + gc->xoffset;
-	points[2].Y = y3 + gc->yoffset;
+	points[0].X = (INT)(x1 + gc->xoffset - 1); // buggy GDI+?
+	points[0].Y = (INT)(y1 + gc->yoffset - 1); // buggy GDI+?
+	points[1].X = (INT)(x2 + gc->xoffset);
+	points[1].Y = (INT)(y2 + gc->yoffset);
+	points[2].X = (INT)(x3 + gc->xoffset);
+	points[2].Y = (INT)(y3 + gc->yoffset);
 
 	g_brushBase->SetColor(XS_RGB(c));
 	g_gcBase->FillPolygon(g_brushBase, points, 3);
@@ -342,7 +342,7 @@ static REAL GetRealSize(xsFontType *font)
 
 	if (font->size < 0) 
 	{// preset font size
-		int idx = -1 - font->size;
+		int idx = -1 - (int)font->size;
 		if (idx >= 0 && idx < XS_FONT_SIZE_COUNT)
 			size = g_fontSizes[idx];
 		else
@@ -548,7 +548,7 @@ void xsDrawImage(xsGraphics *gc, xsImage *img, float x, float y)
 		XS_ERROR("xsDrawImage failed.");
 		return;
 	}
-	g_gcBase->DrawImage((Bitmap *)img->object, Point(x + gc->xoffset, y + gc->yoffset));
+	g_gcBase->DrawImage((Bitmap *)img->object, Point((INT)(x + gc->xoffset), (INT)(y + gc->yoffset)));
 }
 
 #define XS_INPUT_CTRL_ID	1020
