+++ pal/win32/src/main.cpp
@@ -18,8 +18,8 @@ extern "C"
 TCHAR g_szTitle[64];
 TCHAR g_szWindowClass[64];
 char g_szAppId[MAX_PATH] = {0};
-int g_nScreenWidth = 240;
-int g_nScreenHeight = 320;
+float g_nScreenWidth = 240;
+float g_nScreenHeight = 320;
 
 extern Bitmap *g_bmpBase;
 extern Graphics *g_gcBase;
